0.9 - Extra Credit 3
Finished the whole question only to realize I forgot to sort for dogs and cats, also that I need to go back and change some past answers haha Also I downloaded a plugin for Visual studio code and amd finding it a far superior app to do this in. Jupyter Notebooks is glitchyyyyyyy.